requirements.txt
anthropic>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
reportlab>=4.0.0
weasyprint>=60.0
streamlit>=1.28.0
